move systemColors in extension
