+++ Snabble/UI/Utilities/Assets+SystemColors.swift
@@ -1,0 +1,176 @@
+//
+//  Assets+SystemColors.swift
+//  Snabble
+//
+//  Created by Andreas Osberghaus on 11.08.22.
+//
+
+import Foundation
+import UIKit
+
+extension Assets {
+    enum Color {
+        static func systemRed(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemRed", domain: domain) ?? .systemRed
+        }
+
+        static func systemGreen(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGreen", domain: domain) ?? .systemGreen
+        }
+
+        static func systemBlue(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemBlue", domain: domain) ?? .systemBlue
+        }
+
+        static func systemOrange(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemOrange", domain: domain) ?? .systemOrange
+        }
+
+        static func systemYellow(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemYellow", domain: domain) ?? .systemYellow
+        }
+
+        static func systemPink(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemPink", domain: domain) ?? .systemPink
+        }
+
+        static func systemPurple(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemPurple", domain: domain) ?? .systemPurple
+        }
+
+        static func systemTeal(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemTeal", domain: domain) ?? .systemTeal
+        }
+
+        static func systemIndigo(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemIndigo", domain: domain) ?? .systemIndigo
+        }
+
+        static func systemBrown(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemBrown", domain: domain) ?? .systemBrown
+        }
+
+        @available(iOS 15.0, *)
+        static func systemMint(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemMint", domain: domain) ?? .systemMint
+        }
+
+        @available(iOS 15.0, *)
+        static func systemCyan(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemCyan", domain: domain) ?? .systemCyan
+        }
+
+        static func systemGray(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray", domain: domain) ?? .systemGray
+        }
+
+        static func systemGray2(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray2", domain: domain) ?? .systemGray2
+        }
+
+        static func systemGray3(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray3", domain: domain) ?? .systemGray3
+        }
+
+        static func systemGray4(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray4", domain: domain) ?? .systemGray4
+        }
+
+        static func systemGray5(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray5", domain: domain) ?? .systemGray5
+        }
+
+        static func systemGray6(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGray6", domain: domain) ?? .systemGray6
+        }
+
+        @available(iOS 15.0, *)
+        static func tintColor(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "tintColor", domain: domain) ?? .tintColor
+        }
+
+        static func label(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "label", domain: domain) ?? .label
+        }
+
+        static func secondaryLabel(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "secondaryLabel", domain: domain) ?? .secondaryLabel
+        }
+
+        static func tertiaryLabel(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "tertiaryLabel", domain: domain) ?? .tertiaryLabel
+        }
+
+        static func quaternaryLabel(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "quaternaryLabel", domain: domain) ?? .quaternaryLabel
+        }
+
+        static func link(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "link", domain: domain) ?? .link
+        }
+
+        static func placeholderText(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "placeholderText", domain: domain) ?? .placeholderText
+        }
+
+        static func separator(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "separator", domain: domain) ?? .separator
+        }
+
+        static func opaqueSeparator(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "opaqueSeparator", domain: domain) ?? .opaqueSeparator
+        }
+
+        static func systemBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemBackground", domain: domain) ?? .systemBackground
+        }
+
+        static func secondarySystemBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "secondarySystemBackground", domain: domain) ?? .secondarySystemBackground
+        }
+
+        static func tertiarySystemBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "tertiarySystemBackground", domain: domain) ?? .tertiarySystemBackground
+        }
+
+        static func systemGroupedBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemGroupedBackground", domain: domain) ?? .systemGroupedBackground
+        }
+
+        static func secondarySystemGroupedBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "secondarySystemGroupedBackground", domain: domain) ?? .secondarySystemGroupedBackground
+        }
+
+        static func tertiarySystemGroupedBackground(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "tertiarySystemGroupedBackground", domain: domain) ?? .tertiarySystemGroupedBackground
+        }
+
+        static func systemFill(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "systemFill", domain: domain) ?? .systemFill
+        }
+
+        static func secondarySystemFill(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "secondarySystemFill", domain: domain) ?? .secondarySystemFill
+        }
+
+        static func tertiarySystemFill(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "tertiarySystemFill", domain: domain) ?? .tertiarySystemFill
+        }
+
+        static func quaternarySystemFill(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "quaternarySystemFill", domain: domain) ?? .quaternarySystemFill
+        }
+
+        static func lightText(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "lightText", domain: domain) ?? .lightText
+        }
+
+        static func darkText(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "darkText", domain: domain) ?? .darkText
+        }
+
+        static func clear(in domain: Any? = domain) -> UIColor? {
+            Assets.color(named: "clear", domain: domain) ?? .clear
+        }
+    }
+}
